+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.0.2 (2021-06-23)
 - Debounce service defined as interface
+- safe_codes configuration parameter
 
 ## v1.0.0 (2021-06-14)
 - Initial version